Restyled user settings page
The user settings page was getting confusing, with three (3) separate
forms with little visual distinction.

This commit adds some simple styling to visually separate the 3 forms.

* Added a bare-bones multi-column grid using flexbox. It isn't a true
  grid layout system, but allows for half- and full-width columns.
* This allows the "main" form (the edit profile form) to take the top,
  full width position.
* The other two forms appear side-by-side as half-width columns beneath.
* Furthermore, the two "smaller" (not really, but they won't be growing)
  forms are places in a hero-box section to further distinguish them
  from the main form.

A lot of this will likely be refactored once a proper css framework is
added to the project, but for now this will do.